Update getChannel function
